moving nmc and predictive functionality from robust to observational
