+++ tests/robust/test_handlers/test_MC_EIF.py
@@ -8,7 +8,7 @@ import torch
 from typing_extensions import ParamSpec
 
 from chirho.robust.handlers.estimators import MonteCarloInfluenceEstimator
-from chirho.robust.handlers.predictive import PredictiveFunctional, PredictiveModel
+from chirho.observational.handlers.predictive import PredictiveFunctional, PredictiveModel
 from chirho.robust.ops import influence_fn
 
 from ..robust_fixtures import SimpleGuide, SimpleModel
